1.0.2:
A couple of optimizations specific to x86-64 have been applied to the
included MD5 routines.

CVS-ID: packages/popa3d/popa3d/md5/md5.c 1.5
CVS-ID: packages/popa3d/popa3d/md5/md5.h 1.4
CVS-ID: packages/popa3d/popa3d/CHANGES 1.11
CVS-ID: packages/popa3d/popa3d/version.c 1.8
CVS-ID: packages/popa3d/build.archive 1.23
CVS-ID: packages/popa3d/popa3d.spec 1.51

+++ packages/popa3d/popa3d/md5/md5.h
@@ -12,7 +12,7 @@
 #define _MD5_H
 
 /* Any 32-bit or wider unsigned integer data type will do */
-typedef unsigned long MD5_u32plus;
+typedef unsigned int MD5_u32plus;
 
 typedef struct {
 	MD5_u32plus lo, hi;
